// _slack-alarm/src/common/types/docs.d.ts
// The contents of this file are auto-generated from swagger.yaml definitions.
// Do not edit directly.

export interface AlarmDb {
    _id: string; // ^[\da-fA-F]{8}-[\da-fA-F]{4}-[\da-fA-F]{4}-[\da-fA-F]{4}-[\da-fA-F]{12}
    name: string;
    alertAt: Date; // date-time
}
export interface ErrorResponse {
    message: string;
}